+++ src/ansys/simai/core/data/global_coefficients_requests.py
@@ -30,6 +30,7 @@ from ansys.simai.core.data.base import (
     ComputableDataModel,
     Directory,
 )
+from ansys.simai.core.utils.numerical import cast_values_to_float
 
 if TYPE_CHECKING:
     import ansys.simai.core.client
@@ -179,7 +180,7 @@ class ComputeGlobalCoefficient(GlobalCoefficientRequest):
             self._set_is_pending()
         elif state == "successful":
             logger.debug(f"{self._classname} id {self.id} set status successful")
-            self._result = data.get("result").get("value")
+            self._result = cast_values_to_float(data.get("result", {}).get("value"))
             self._set_is_over()
         elif state in ERROR_STATES:
             error_message = f"Computation of global coefficient {target.get('formula')} failed with {data.get('reason', 'UNKNOWN ERROR')}"